% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional-functions.R
\name{AmpFromLocation}
\alias{AmpFromLocation}
\title{Get amplitude of seasonal cycle from location}
\usage{
AmpFromLocation(
  longitude,
  latitude,
  depth.upr,
  depth.lwr,
  proxy.type = c("degC", psem:::proxy.types)
)
}
\arguments{
\item{longitude}{Longitude}

\item{latitude}{Latitude}

\item{depth.upr}{Upper end of depth range, as negative metres below sealevel}

\item{depth.lwr}{Lower end of depth range, as negative metres below sealevel}

\item{proxy.type}{Proxy type, decC or d18O}
}
\value{
A list
}
\description{
Get amplitude of seasonal cycle from location
}
\examples{
AmpFromLocation(120.5, 0.5, 0, -120, proxy.type = "d18O")
}
