% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec-model.R
\name{ClimPowerFunction}
\alias{ClimPowerFunction}
\title{Power function for climate}
\usage{
ClimPowerFunction(nu, a, b)
}
\arguments{
\item{nu}{frequency}

\item{a}{scaling of the power function}

\item{b}{exponent of power function}
}
\value{
vector of power
}
\description{
Power function for climate
}
\examples{
ClimPowerFunction(1/10, 0.125, 1)
}
