% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec-model.R
\name{ExpectedCorrelation}
\alias{ExpectedCorrelation}
\title{Expected correlation between replicate proxy records and between a proxy record and the true climate}
\usage{
ExpectedCorrelation(pes, spec.pars = NULL)
}
\arguments{
\item{pes}{Object of class proxy.error.spec, e.g. output from \link{ProxyErrorSpectrum}}

\item{spec.pars}{Parameters of the proxy error spectrum, these are taken
from proxy.error.spec if it is a proxy.error.spec object. Can be passed here to
allow calculation on compatible none proxy.error.spec objects.}
}
\value{
a data.frame / tibble
}
\description{
Expected correlation between replicate proxy records and between a proxy record and the true climate
}
\examples{
spec.pars <- GetSpecPars("Mg_Ca", tau_b = 1000 * 10 / 2, sigma.meas = 1)
spec.obj <- do.call(ProxyErrorSpectrum, spec.pars)

exp.corr <- ExpectedCorrelation(spec.obj)
plot(exp.corr, type = "l", log = "x")
}
