% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec-model.R
\name{GetNu}
\alias{GetNu}
\title{Get the discrete frequencies of the power spectrum of a regularly sampled
 timeseries.}
\usage{
GetNu(T, delta_t)
}
\arguments{
\item{T}{length of a finite time series, e.g. 1e04}

\item{delta_t}{sampling frequency of the sediment core / climate timeseries}
}
\value{
a vector of discrete frequencies
}
\description{
Get the discrete frequencies of the power spectrum of a regularly sampled
 timeseries.
}
\examples{
GetNu(1e04, 100)
}
