% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec-model.R
\name{GetProxyError}
\alias{GetProxyError}
\title{Get error from proxy.error.var object}
\usage{
GetProxyError(
  var.obj,
  timescale,
  exclude = NULL,
  include.f.zero = TRUE,
  format = "long"
)
}
\arguments{
\item{var.obj}{timescale dependent variance object from \link{IntegrateErrorSpectra}}

\item{timescale}{the temporal resolution at which the proxy values are being interpreted}

\item{exclude}{character vector of error components to exclude from the total}

\item{include.f.zero}{include or exclude power at nu = 0}

\item{format}{format of the output}
}
\value{
1 row dataframe with 1 SD error components
}
\description{
Get error from proxy.error.var object
}
\examples{
spec.pars <- GetSpecPars("Mg_Ca", tau_p = 1 / 12, phi_c = 0, seas.amp = 4, T = 100 * 101)
spec.obj <- do.call(ProxyErrorSpectrum, spec.pars)
PlotSpecError(spec.obj)
var.obj <- IntegrateErrorSpectra(spec.obj)
PlotTSDVariance(var.obj)
GetProxyError(var.obj, timescale = 100)
}
