% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec-model.R
\name{GetSpecPars}
\alias{GetSpecPars}
\title{Create list of required parameters for ProxyErrorSpectrum}
\usage{
GetSpecPars(proxy.type, ...)
}
\arguments{
\item{proxy.type}{"Mg_Ca", "T_deg_Mg_Ca", "T_deg_Uk37", "Uk37" or "d18O"}

\item{...}{Arguments to ProxyErrorSpectrum, e.g. delta_t = 1000}
}
\value{
A list of arguments to \code{\link{ProxyErrorSpectrum}}
}
\description{
Returns a list of arguments to \code{\link{ProxyErrorSpectrum}}
 where arguments in the call replace the defaults. Default values are returned
 for all arguments not in the call.
}
\examples{
\dontrun{
GetSpecPars()
}
GetSpecPars(proxy.type = "Mg_Ca")
GetSpecPars(proxy.type = "Uk37")
do.call(ProxyErrorSpectrum, GetSpecPars("Mg_Ca"))
}
