% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-spectrum.R
\name{ModelSpectrum}
\alias{ModelSpectrum}
\title{Create a model SST spectrum based on the observed spectra and a
  powerlaw spectrum}
\usage{
ModelSpectrum(
  freq = NULL,
  latitude,
  spectable = psem:::spectable,
  beta = 1,
  variable = c("temperature", "d18O", "T_deg_Mg_Ca", "T_deg_Uk37", "d18O"),
  freq.match.lower = 0.03,
  freq.match.upper = 0.1,
  bPlot = FALSE
)
}
\arguments{
\item{freq}{desired frequency vector}

\item{latitude}{latitude in degN}

\item{spectable}{lookup table with spectra by latitude
list(spec[iLat,iFreq],lat[iLat],freq[iFreq])}

\item{beta}{desired scaling}

\item{variable}{climate variable for which to return the power spectrum,
currently temperature or d18O}

\item{freq.match.lower}{lower frequency boundary at which the variance is
matched and the spectrum is switched from instrumental to powerlaw (default
0.03 as lowest reliable estimate from the instrumental record given the
detrending and the multitaper bias)}

\item{freq.match.upper}{higher frequency boundary at which the variance is
matched (default 0.1 = decadal)}

\item{bPlot}{TRUE = provide diagnostic plot}
}
\value{
model spectrum list(spec,freq)
}
\description{
Create a model SST spectrum based on the observed spectra and a
  powerlaw spectrum
}
\examples{
ModelSpectrum(latitude = 20,beta=1, bPlot=TRUE)
ModelSpectrum(freq = GetNu(T = 10000, delta_t = 100), latitude = 20, beta=1, bPlot=FALSE)
ModelSpectrum(freq = c(0.1, GetNu(T = 10000, delta_t = 100), 0.1), latitude = 20, beta=1, bPlot=FALSE)
ModelSpectrum(freq = c(0.6, 1), latitude = 20, beta=1, bPlot=FALSE)
}
\author{
Thomas Laepple
}
