% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec-model.R
\name{OrbitalError}
\alias{OrbitalError}
\title{Get the orbital error components}
\usage{
OrbitalError(
  nu = NULL,
  tau_s,
  tau_b,
  tau_p,
  delta_t,
  T,
  sig.sq_c,
  phi_c,
  delta_phi_c,
  nu_c,
  sig.sq_a,
  phi_a,
  nu_a,
  N
)
}
\arguments{
\item{nu}{frequency}

\item{tau_s}{sediment slice thickness in years (layer.width / sedimentation rate)}

\item{tau_b}{timescale of bioturbation (bioturbation depth / sedimentation rate) (L/sr)}

\item{tau_p}{length of proxy carrier "growth season" (< 12 months)}

\item{delta_t}{sampling frequency of the sediment core / climate timeseries}

\item{T}{length of a finite time series, e.g. 1e04}

\item{sig.sq_c}{variance of the seasonal cycle}

\item{phi_c}{phase of carrier growth season relative to the maximum of the seasonal cycle}

\item{nu_c}{1/1 (yr) = frequency of the seasonal cycle}

\item{sig.sq_a}{variance of precessionary amplitude modulation}

\item{phi_a}{phase of precessionary amplitude modulation relative to centre of finite timeseries of length T}

\item{nu_a}{1/tau_a = frequency of the orbital variation, e.g. precession}

\item{N}{number of signal carriers (e.g. individual foraminifera)}
}
\value{
A list with the frequency domain and time domain error components
}
\description{
Get the orbital error components
}
\examples{
spec.pars <- GetSpecPars("Mg_Ca", phi_c = pi, delta_phi_c = pi, sig.sq_a = 0.1, tau_p = 1/12)
do.call(OrbitalError, spec.pars[names(spec.pars) \%in\% names(formals(OrbitalError))])
}
