% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec-model.R
\name{OrderStages}
\alias{OrderStages}
\title{Order the stages/components of the error}
\usage{
OrderStages(vec, rev = FALSE)
}
\arguments{
\item{vec}{vector of names of error stages/components}
}
\value{
An ordered factor
}
\description{

}
\details{

}
