% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec-model.R
\name{plot.proxy.error.spec}
\alias{plot.proxy.error.spec}
\title{Plot method for error spec}
\usage{
\method{plot}{proxy.error.spec}(x, ...)
}
\arguments{
\item{x}{proxy.error.spec object}

\item{...}{other parameters passed to PlotSpecError}
}
\description{
Plot method for error spec
}
