% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-spectrum.R
\name{PlotModelSpectrum}
\alias{PlotModelSpectrum}
\title{Plot a spliced empirical / power-law climate spectrum}
\usage{
PlotModelSpectrum(model.spec)
}
\arguments{
\item{model.spec}{Output from ModelSpectrum}
}
\description{
Plot a spliced empirical / power-law climate spectrum
}
\examples{
clim.spec <- ModelSpectrum(
  freq = NULL,
  latitude = 20,
  variable = "temperature",
   beta = 1)

PlotModelSpectrum(clim.spec)
}
