% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional-functions.R
\name{PlotProxySeasonality}
\alias{PlotProxySeasonality}
\title{PlotProxySeasonality}
\usage{
PlotProxySeasonality(phi_c, tau_p, res = 365)
}
\arguments{
\item{phi_c}{phase of carrier growth season relative to the maximum of the seasonal cycle}

\item{tau_p}{length of proxy carrier "growth season" (< 12 months)}

\item{res}{resolution}
}
\value{

}
\description{
PlotProxySeasonality
}
\examples{
PlotProxySeasonality(phi_c = 0, tau_p = 1/12, res = 365)
PlotProxySeasonality(phi_c = 0, tau_p = 6/12, res = 365)
PlotProxySeasonality(phi_c = -pi, tau_p = 3/12, res = 365)
PlotProxySeasonality(phi_c = -pi, tau_p = 1/12, res = 12)
PlotProxySeasonality(phi_c = pi, tau_p = 2/12, res = 12)
}
