% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec-model.R
\name{PlotSpecError}
\alias{PlotSpecError}
\title{Plot a proxy error spectrum}
\usage{
PlotSpecError(pes, show.low.power.panel = FALSE)
}
\arguments{
\item{pes}{Object of class proxy.error.spec, e.g. output from ProxyErrorSpectrum}

\item{show.low.power.panel}{Show components with very low power in an
additional panel (otherwise they are excluded)}
}
\value{
a ggplot object
}
\description{
Plot a proxy error spectrum
}
\examples{
spec.pars <- psem::GetSpecPars("Mg_Ca", T = 1e04)
spec.obj <- do.call(psem::ProxyErrorSpectrum, spec.pars)
PlotSpecError(spec.obj)
}
