% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec-model.R
\name{PlotTSDVariance}
\alias{PlotTSDVariance}
\title{Plot Time-scale Dependent Variance Object.}
\usage{
PlotTSDVariance(
  var.obj,
  components = c("Bioturbation", "Seasonal.bias", "Seasonal.bias.unc.",
    "Calibration.unc.", "Aliasing.seasonal", "Aliasing.stochastic",
    "Individual.variation", "Meas.error"),
  include = TRUE,
  include.constant.errors = FALSE,
  units = c("degC", "d18O"),
  thin.var.obj = TRUE
)
}
\arguments{
\item{var.obj}{Object of class proxy.error.var, e.g. output from \link{IntegrateErrorSpectra}}

\item{components}{Which error components to include or exclude}

\item{include}{Include or exclude components named in \code{components}}

\item{include.constant.errors}{Include the constant (nu = 0) error components}

\item{units}{Are the units degC or d18O, used for axis labelling}

\item{thin.var.obj}{If TRUE and the variance object contains more than 1000 frequencies
these are thinned to 1000 frequencies equally spaced on a log-scale}
}
\value{
ggplot object
}
\description{
Plot Time-scale Dependent Variance Object.
}
\examples{
spec.pars <- GetSpecPars("Mg_Ca", delta_phi_c = pi)
spec.obj <- do.call(ProxyErrorSpectrum, spec.pars)

var.obj <- IntegrateErrorSpectra(spec.obj)
PlotTSDVariance(var.obj, units = "degC")
PlotTSDVariance(var.obj, units = "d18O")
PlotTSDVariance(var.obj, include = FALSE, c("Seasonal.bias", "Calibration.unc."))
PlotTSDVariance(var.obj, include.constant.error = TRUE)
}
