% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec-model.R
\name{ProxyErrorSpectrum}
\alias{ProxyErrorSpectrum}
\title{Power Spectral Density of the Error in a Sediment Archived Proxy Timeseries}
\usage{
ProxyErrorSpectrum(
  nu = NULL,
  tau_s,
  tau_b,
  tau_p,
  tau_r,
  T,
  delta_t,
  N,
  n.k,
  clim.spec.fun,
  clim.spec.fun.args,
  sig.sq_a,
  sig.sq_c,
  nu_a,
  nu_c,
  phi_a,
  phi_c,
  delta_phi_c,
  sigma.meas,
  sigma.ind,
  sigma.cal,
  n.nu.prime = 1000,
  ...
)
}
\arguments{
\item{nu}{frequency}

\item{tau_s}{sediment slice thickness in years (layer.width / sedimentation rate)}

\item{tau_b}{timescale of bioturbation (bioturbation depth / sedimentation rate) (L/sr)}

\item{tau_p}{length of proxy carrier "growth season" (< 12 months)}

\item{tau_r}{width of moving average filter that represents the "interpreted" resolution of the timeseries}

\item{T}{length of a finite time series, e.g. 1e04}

\item{delta_t}{sampling frequency of the sediment core / climate timeseries}

\item{N}{number of signal carriers (e.g. individual foraminifera)}

\item{clim.spec.fun}{a function to return climate power spectral density as a function of frequency, nu}

\item{clim.spec.fun.args}{arguments to the named climate power spectrum function}

\item{sig.sq_a}{variance of precessionary amplitude modulation}

\item{sig.sq_c}{variance of the seasonal cycle}

\item{nu_a}{1/tau_a = frequency of the orbital variation, e.g. precession}

\item{nu_c}{1/1 (yr) = frequency of the seasonal cycle}

\item{phi_a}{phase of precessionary amplitude modulation relative to centre of finite timeseries of length T}

\item{phi_c}{phase of carrier growth season relative to the maximum of the seasonal cycle}

\item{sigma.meas}{the standard deviation of the per sample measurement error}

\item{sigma.ind}{the standard deviation of error between individuals (e.g.
foraminifera) e.g. due to "vital effects" or calcification depth}

\item{n.nu.prime}{number of discrete frequencies at which to evaluate the PSD of the error}

\item{tau_a}{period of orbital cycle, e.g. 23e03 yrs}
}
\value{
a dataframe of frequencies and spectral power
}
\description{
Power Spectral Density of the Error in a Sediment Archived Proxy Timeseries
}
\examples{
spec.pars <- GetSpecPars("Mg_Ca", tau_p = 4 / 12, delta_phi_c = pi)
spec.obj <- do.call(ProxyErrorSpectrum, spec.pars)
PlotSpecError(spec.obj)
}
