% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psem.R
\docType{package}
\name{psem}
\alias{psem}
\title{psem: Proxy Spectral Error Model}
\description{
An R package to implement the PSEM spectral error model for proxy records.
}
\details{
PSEM is described in the discussion papers:

Estimating the timescale-dependent uncertainty of paleoclimate records—a 
spectral approach. Part I: Theoretical concept.

Estimating the timescale-dependent uncertainty of paleoclimate records—a 
spectral approach. Part II: Application and interpretation.
}
