% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec-model.R
\name{reverselog_trans}
\alias{reverselog_trans}
\title{Title}
\source{
https://stackoverflow.com/a/11054781/1198125
}
\usage{
reverselog_trans(base = exp(1))
}
\arguments{
\item{base}{base for the log transformation, Default: exp(1)}
}
\description{
Custom axis transformation function
}
\seealso{
\code{\link[scales]{trans_new}},\code{\link[scales]{log_breaks}}
}
