% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec-model.R
\name{rhoCI}
\alias{rhoCI}
\title{Confidence interval of a correlation coefficient}
\usage{
rhoCI(rho, n, alpha = 0.05)
}
\arguments{
\item{rho}{correlation coefficient}

\item{n}{number of data points used to estimate correlation coefficient}

\item{alpha}{confidence level}
}
\value{
a dataframe
}
\description{
Confidence interval of a correlation coefficient
}
\examples{
rhoCI(rho = seq(0.1, 0.9, 0.1), n = c(100))
}
