% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec-model.R
\name{S_E}
\alias{S_E}
\title{Get error components for stochastic climate signal}
\usage{
S_E(
  nu,
  tau_s,
  tau_b,
  tau_r,
  T,
  delta_t,
  N,
  n.k,
  clim.spec.fun,
  clim.spec.fun.args,
  n.nu.prime = 1000,
  plot.integral = FALSE
)
}
\arguments{
\item{nu}{frequency}

\item{tau_s}{sediment slice thickness in years (layer.width / sedimentation
rate)}

\item{tau_b}{timescale of bioturbation (bioturbation depth / sedimentation
rate) (L/sr)}

\item{tau_r}{width of moving average filter that represents the "interpreted"
resolution of the timeseries}

\item{T}{length of a finite time series, e.g. 1e04}

\item{delta_t}{sampling frequency of the sediment core / climate timeseries}

\item{N}{number of signal carriers (e.g. individual foraminifera)}

\item{n.k}{the number of aliasers used when estimating the error spectrum of 
the stochastic climate signal. Defaults to 15, does not normally need to be 
adjusted.}

\item{clim.spec.fun}{a function to return climate power spectral density as a
function of frequency, nu}

\item{clim.spec.fun.args}{arguments to the named climate power spectrum
function}

\item{n.nu.prime}{number of discrete frequencies at which to evaluate the PSD
of the error}

\item{plot.integral}{Plot the PSD of the error of the bioturbated climate
timeseries. This PSD is numerically integrated to give the variance for the
aliasing of climate variation.}
}
\value{
list of frequencies and error spectrum components
}
\description{
Get error components for stochastic climate signal
}
\examples{
spec.pars <- GetSpecPars("Mg_Ca", phi_c = pi, delta_phi_c = pi, sig.sq_a = 0.1, tau_p = 1/12)
spec.pars$nu <- GetNu(spec.pars$T, spec.pars$delta_t)
do.call(S_E, spec.pars[names(spec.pars) \%in\% names(formals(S_E))])
}
\seealso{
\code{\link{ProxyErrorSpectrum}}
}
