% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec-model.R
\name{TF.f_bs}
\alias{TF.f_bs}
\title{Bioturbation Transfer Function.}
\usage{
TF.f_bs(nu, tau_b, tau_s)
}
\arguments{
\item{nu}{frequency}

\item{tau_b}{timescale of bioturbation (bioturbation depth / sedimentation
rate) (L/sr)}

\item{tau_s}{sediment slice thickness in years (layer.width / sedimentation
rate)}
}
\value{
a vector
}
\description{
Fourier transform of bioturbation and sediment slice thickness filter.
}
\examples{
nu = seq(1/10000, 1/50, 1/10000)
tf <- TF.f_bs(nu = nu, tau_b = 1000*10/50, tau_s = 1000*1/50)
plot(nu, abs(tf)^2, type = "l", log = "xy")
}
