% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec-model.R
\name{VarSine}
\alias{VarSine}
\title{Variance of a Sine wave}
\usage{
VarSine(full.amp)
}
\arguments{
\item{full.amp}{Peak to peak amplitude of a sine wave}
}
\value{
numeric, the variance of a sine wave
}
\description{
Variance of a Sine wave
}
\examples{
x <- sin(seq(0, 2 * pi, length.out = 1000))
plot(x, type = "l")
var(x)
VarSine(2)
x5 <- x * 5
var(x5)
VarSine(diff(range(x5)))
}
