% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec-model.R
\name{WhiteNoise}
\alias{WhiteNoise}
\title{Get spectrum of white noise error components}
\usage{
WhiteNoise(sigma.meas, sigma.ind, N, delta_t, T)
}
\arguments{
\item{sigma.meas}{the standard deviation of the per sample measurement error}

\item{sigma.ind}{the standard deviation of error between individuals (e.g.
foraminifera) e.g. due to "vital effects" or calcification depth}

\item{N}{number of signal carriers (e.g. individual foraminifera)}

\item{delta_t}{sampling frequency of the sediment core / climate timeseries}

\item{T}{length of a finite time series, e.g. 1e04}
}
\description{
Get spectrum of white noise error components
}
\keyword{internal}
