% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec-model.R
\name{sinc}
\alias{sinc}
\title{Normalized sinc function}
\usage{
sinc(x, normalized = TRUE)
}
\arguments{
\item{x}{numeric}

\item{normalized}{logical, return the normalized or non-normalized sinc
function, defaults to TRUE}
}
\description{
Return the normalized or non-normalized sinc function. Returns 1
for x == 0 
sinc(x) = sin(pi * x) / (pi * x)
sinc(x) = sin(x) / (x)
}
\examples{
x <- seq(-10, 10, length.out = 1000)
plot(x, sinc(x), type = "l")
lines(x, sinc(x)^2, col = "Red")
abline(h = 0)

x <- seq(-10, 10, length.out = 1000)
plot(x, sinc(x, normalized = FALSE), type = "l")
lines(x, sinc(x, normalized = FALSE)^2, col = "Red")
abline(h = 0)
}
