% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeslice-error.R
\name{SortForFFT}
\alias{SortForFFT}
\title{Sort a power spectrum into the correct order for fft(., inverse = T)}
\usage{
SortForFFT(spec.obj)
}
\arguments{
\item{spec.obj}{proxy.error.spec object, list (or dataframe) of frequencies (nu)
and power (spec)}
}
\value{
a dataframe
}
\description{
Sort a power spectrum into the correct order for fft(., inverse = T)
}
\examples{
nu <- seq(-1/2, 1/2, 1/10)
dat <- data.frame(nu = nu, spec = 0.1*nu^1)
SortForFFT(dat)
}
